Application is now split in a blueprint for authentication and admin and a blueprint for the main application
