+++ devices/hmi/software/app/templates/auth_base.html
@@ -1,13 +1,13 @@
 <!doctype html>
-<title>{% block title %}{% endblock %} - Flaskr</title>
+<title>{% block title %}{% endblock %} LICSTER</title>
 <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
 <nav>
   <ul>
     {% if g.user %}
       <li><span>{{ g.user['username'] }}</span>
-      <li><a href="{{ url_for('views.logout') }}">Log Out</a>
+      <li><a href="{{ url_for('auths.logout') }}">Log Out</a>
     {% else %}
-      <li><a href="{{ url_for('views.login') }}">Log In</a>
+      <li><a href="{{ url_for('auths.login') }}">Log In</a>
     {% endif %}
   </ul>
 </nav>
